requirements.txt
Django==3.2.12
Pillow==8.4.0
django-widget-tweaks==1.4.11
django-ckeditor
